Alarms but they actually work
Finished up the System side of alarms,
Alarms UI to work on next.

+++ RP2040-Touch-LCD-1.28/previewNotif.h
@@ -3,7 +3,7 @@
 extern bool inTransition;
 extern bool pauseRender;
 extern bool tap;
-extern int CurTime;
+extern unsigned long int CurTime;
 extern int tappedNotif;
 extern UWORD *BlackImage;
 extern std::list<std::string> backgroundApps;
